fix: update to PF6 betas + some style tweaks
