+++ packages/module/package.json
@@ -45,7 +45,7 @@
     "docs:copy": "mkdir -p dist/patternfly-docs && cp -R patternfly-docs/content/extensions/quick-starts dist/patternfly-docs"
   },
   "peerDependencies": {
-    "@patternfly/react-core": "6.0.0-alpha.61",
+    "@patternfly/react-core": "6.0.0-alpha.94",
     "react": ">=18.0.0",
     "react-dom": ">=18.0.0",
     "showdown": ">=2.1.0"
@@ -55,14 +55,14 @@
     "history": "^5.0.0"
   },
   "devDependencies": {
-    "@patternfly/documentation-framework": "6.0.0-alpha.64",
-    "@patternfly/patternfly": "6.0.0-alpha.190",
-    "@patternfly/react-icons": "6.0.0-alpha.31",
-    "@patternfly/react-styles": "6.0.0-alpha.30",
-    "@patternfly/react-tokens": "6.0.0-alpha.30",
-    "@patternfly/react-table": "6.0.0-alpha.89",
-    "@patternfly/react-code-editor": "6.0.0-alpha.88",
-    "@patternfly/react-core": "6.0.0-alpha.88",
+    "@patternfly/documentation-framework": "6.0.0-alpha.69",
+    "@patternfly/patternfly": "6.0.0-alpha.205",
+    "@patternfly/react-icons": "6.0.0-alpha.34",
+    "@patternfly/react-styles": "6.0.0-alpha.33",
+    "@patternfly/react-tokens": "6.0.0-alpha.33",
+    "@patternfly/react-table": "6.0.0-alpha.95",
+    "@patternfly/react-code-editor": "6.0.0-alpha.94",
+    "@patternfly/react-core": "6.0.0-alpha.94",
     "@rollup/plugin-commonjs": "^17.0.0",
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-node-resolve": "^11.1.0",
