+++ packages/vscode/package.json
@@ -61,7 +61,7 @@
   },
   "dependencies": {
     "@cloudmosaic/quickstarts": "^0.0.6",
-    "@patternfly/patternfly": "6.0.0-alpha.190",
+    "@patternfly/patternfly": "6.0.0-alpha.205",
     "i18next-parser": "^9.0.0",
     "js-base64": "^3.6.0",
     "js-yaml": "^4.0.0",
